refactor: move shared utils into a single file
