+++ packages/axelar-local-dev-cosmos/src/__tests__/RemoteAccountCreation.spec.ts
@@ -4,132 +4,8 @@ import { Contract } from 'ethers';
 import { ethers } from 'hardhat';
 import '@nomicfoundation/hardhat-chai-matchers';
 import { HardhatEthersSigner } from '@nomicfoundation/hardhat-ethers/signers';
-import { keccak256, stringToHex, toBytes, encodeAbiParameters } from 'viem';
-import { approveMessage } from './lib/utils';
-
-/**
- * Compute CREATE2 address for RemoteAccount
- * Salt includes routerAddress to prevent front-running attacks
- */
-const computeRemoteAccountAddress = async (
-    factoryAddress: string,
-    portfolioLCA: string,
-    routerAddress: string,
-) => {
-    const salt = ethers.solidityPackedKeccak256(
-        ['string', 'address'],
-        [portfolioLCA, routerAddress],
-    );
-
-    const RemoteAccountFactory = await ethers.getContractFactory('RemoteAccount');
-    const constructorArgs = ethers.AbiCoder.defaultAbiCoder().encode(['string'], [portfolioLCA]);
-    const initCode = ethers.solidityPacked(
-        ['bytes', 'bytes'],
-        [RemoteAccountFactory.bytecode, constructorArgs],
-    );
-    const initCodeHash = ethers.keccak256(initCode);
-
-    return ethers.getCreate2Address(factoryAddress, salt, initCodeHash) as `0x${string}`;
-};
-
-interface ContractCall {
-    target: `0x${string}`;
-    data: `0x${string}`;
-}
-
-interface DepositPermit {
-    tokenOwner: `0x${string}`;
-    permit: {
-        permitted: {
-            token: `0x${string}`;
-            amount: bigint;
-        };
-        nonce: bigint;
-        deadline: bigint;
-    };
-    witness: `0x${string}`;
-    witnessTypeString: string;
-    signature: `0x${string}`;
-}
-
-interface RouterPayloadParams {
-    id: string;
-    portfolioLCA: string;
-    remoteAccountAddress: `0x${string}`;
-    provideAccount: boolean;
-    depositPermit?: DepositPermit[];
-    multiCalls?: ContractCall[];
-}
-
-/**
- * Encode RouterPayload for PortfolioRouter
- */
-const encodeRouterPayload = ({
-    id,
-    portfolioLCA,
-    remoteAccountAddress,
-    provideAccount,
-    depositPermit = [],
-    multiCalls = [],
-}: RouterPayloadParams) => {
-    return encodeAbiParameters(
-        [
-            {
-                type: 'tuple',
-                components: [
-                    { name: 'id', type: 'string' },
-                    { name: 'portfolioLCA', type: 'string' },
-                    { name: 'remoteAccountAddress', type: 'address' },
-                    { name: 'provideAccount', type: 'bool' },
-                    {
-                        name: 'depositPermit',
-                        type: 'tuple[]',
-                        components: [
-                            { name: 'tokenOwner', type: 'address' },
-                            {
-                                name: 'permit',
-                                type: 'tuple',
-                                components: [
-                                    {
-                                        name: 'permitted',
-                                        type: 'tuple',
-                                        components: [
-                                            { name: 'token', type: 'address' },
-                                            { name: 'amount', type: 'uint256' },
-                                        ],
-                                    },
-                                    { name: 'nonce', type: 'uint256' },
-                                    { name: 'deadline', type: 'uint256' },
-                                ],
-                            },
-                            { name: 'witness', type: 'bytes32' },
-                            { name: 'witnessTypeString', type: 'string' },
-                            { name: 'signature', type: 'bytes' },
-                        ],
-                    },
-                    {
-                        name: 'multiCalls',
-                        type: 'tuple[]',
-                        components: [
-                            { name: 'target', type: 'address' },
-                            { name: 'data', type: 'bytes' },
-                        ],
-                    },
-                ],
-            },
-        ],
-        [
-            {
-                id,
-                portfolioLCA,
-                remoteAccountAddress,
-                provideAccount,
-                depositPermit,
-                multiCalls,
-            },
-        ],
-    );
-};
+import { keccak256, stringToHex, toBytes } from 'viem';
+import { approveMessage, computeRemoteAccountAddress, encodeRouterPayload } from './lib/utils';
 
 describe('PortfolioRouter - RemoteAccountCreation', () => {
     let owner: HardhatEthersSigner, addr1: HardhatEthersSigner;
